refactor: extract logic for reusing
